Add files via upload
To better understand heart disease and the measures that could be taken to combat it, the Framingham Heart Study (FHS) was established in the late 1940's.
It was a joint project of Boston University and the National Heart, Lung, and Blood Institute (NHLBI).
A large cohort of initially healthy patients between the age group 30 and 59 in the city of Framingham, Massachusetts were tracked for a period of 20 years, to better understand cardiovascular disease. The study was conducted with an initial cohort of 5209 patients.
The aim of the study was to enroll people free of the disease, and see who developed the disease in the next 20 years.
How was this done?
Every two years, the participants would have to report to a testing center, where an examination was conducted. The patients were examined and their health information was updated.
They were also given questionnaires to fill up, in which they updated behavioral information, such as exercise or smoking habits.
The data collected from this study allowed for a better understanding of the risk factors of heart disease. Medical interventions then took place based on the findings of the FHS.
Important Milestones
Some important milestones of the Framingham Heart Study:
Smoking was found to increase the risk of CHD (1960)
Cholesterol and high BP increased the risk of CHD (1961)
Physical activity decreased the risk of CHD (1967)
High levels of HDL cholesterol was found to increase the risk of CHD (1988)
The lifetime risk of developing CHD was higher in men than in women (1999)
Obesity is a risk factor for heart failure (2002)
Now, we will attempt to build a decision tree classifier in Python that can predict the ten year risk of a patient developing CHD, given certain risk factors.
I will be using a public dataset that can be obtained from the BioLINCC website or else u can download the dataset file from this repository.